### 2021-09-14 #3729 @whedon
[PRE REVIEW]: SNEWPY: A Data Pipeline from Supernova Simulations to Neutrino Signals
**Submitting author:** @JostMigenda (<a href="http://orcid.org/0000-0002-5350-8049">Jost Migenda</a>)
**Repository:** <a href="https://github.com/SNEWS2/snewpy" target ="_blank">https://github.com/SNEWS2/snewpy</a>
**Version:** v1.1b3
