### 2021-09-14 #3729 @JostMigenda
A few suggestions for potential reviewers: **apizzuto**, **bradkav** (who has [previously reviewed a JOSS paper I submitted](https://github.com/openjournals/joss-reviews/issues/2877)), **peterdenton** (not on the JOSS reviewer list, but [interested in neutrino codes](https://peterdenton.github.io/Code/index.html)) and **rafaelab**. (And, if we extend the search a bit, also **kostunin**, **KAClough** or people in the list of potential reviewers who include "astrophysics" in their topics.)

Note also that, later this week, we plan to submit a separate paper that’s more focussed on the underlying physics to AAS Journals and would like to make use of [the AAS/JOSS collaboration](https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing).
