### 2021-09-15 #3729 @Kevin-Mattheus-Moerman
@JostMigenda thanks for this submission. As you can see from my comment, many of our editors in this domain are currently pre-ocupied with other submissions as JOSS is processing a bit of a backlog at the moment. As such I have listed this work as `waitlisted`, which means an editor will get to it once they are available. Thank you for your patience.
