### 2021-09-15 #3729 @JostMigenda
Thanks for volunteering—and yes, happy to wait a few days longer if it helps you stick to a sustainable workload!

One first question to the editor right away: As is common in astro/particle physics, this work was done not just by a small number of people but as part of a larger collaboration (in this case the SNEWS Collaboration). Is it possible to include the collaboration name after the author list in a JOSS paper? (E.g. like the AASTeX class, which has a [special `\collaboration{}{}` command](https://journals.aas.org/aastex-v6-3-author-guide/#title_author) for that.)
