### 2021-09-26 #3729 @dfm
@JostMigenda: I'm working on finding reviewers now and I'll keep you posted on that!

> One first question to the editor right away: As is common in astro/particle physics, this work was done not just by a small number of people but as part of a larger collaboration (in this case the SNEWS Collaboration). Is it possible to include the collaboration name after the author list in a JOSS paper? (E.g. like the AASTeX class, which has a special \collaboration{}{} command for that.)

Sorry about the delay on this response, but I actually had to do a little digging. We don't currently have a special command for this and our usual recommendation for this is to just add the collaboration as an author. You'll need to include an affiliation, so perhaps choose the primary institution or just label it "no affiliation".
