### 2021-09-27 #3729 @apizzuto
Hi all, thanks for the submission!
The [JOSS COI policy](https://github.com/openjournals/joss/blob/master/COI.md) requests that reviewers disclose potential COIs. I am a member of the IceCube Collaboration and there are at least a few folks on the author list of this paper that are also members (Segev BenZvi, Spencer Griswold, and I may have missed others). The COI policy says that for large collaborations this is often fine as long as we did not work together.
I believe that I am enough removed to review the contribution without bias, and I am happy to still review this project if the authors believe I can review this impartially and are comfortable waiving this COI, but I am happy to defer to the authors.
