### 2021-09-27 #3729 @danielskatz
FYI, the waived COI has to be approved by the editor, not by the authors.  (not that I see any problem, but please wait for @dfm to decide.)
