### 2021-09-27 #3729 @dfm
Thanks @danielskatz!

Yes, given the scope of this potential COI and @JostMigenda's response, I am happy to waive this case and proceed with the review. Thanks again @apizzuto for being transparent. I'll get the review started in a new thread and we'll move over there for the rest of the process. Thanks again, everyone!
