### 2021-09-30 8c8c2cf
new version of snewpy_to_snewpdag script working with last released version
