+++ models/Nakazato_2013/README.md
@@ -1,5 +1,5 @@
-CCSN neutrino models from Nakazato et al., 2013 and 2015. Data are public and
-taken from [their website](http://asphwww.ph.noda.tus.ac.jp/snn/).
+CCSN neutrino models from Nakazato et al., 2013, 2015 and 2021. Data are available publicly
+on [their website](http://asphwww.ph.noda.tus.ac.jp/snn/) and were reprocessed for SNEWPY.
 
 The citation for use of the database is: *Supernova Neutrino Light Curves and
 Spectra for Various Progenitor Stars: From Core Collapse to Proto-neutron Star
@@ -11,3 +11,8 @@ If the BH model with LS220 EOS is used, the citation is: *Spectrum of the
 Supernova Relic Neutrino Background and Metallicity Evolution of Galaxies*, 
 K. Nakazato, E. Mochida, Y. Niino, and H. Suzuki, 
 [Astrophys. J. 804 (2015) 75](http://dx.doi.org/10.1088/0004-637X/804/1/75), [arXiv:1503.01236](http://arxiv.org/abs/1503.01236).
+
+For the BH model with LS220 EOS, the citation is: *Numerical Study of Stellar Core
+Collapse and Neutrino Emission Using the Nuclear Equation of State Obtained by the Variational Method*,
+K. Nakazato, K. Sumiyoshi, and H. Togashi, 
+[Publ. Astron. Soc. Jpn. 73 (2021) 639-651](https://doi.org/10.1093/pasj/psab026), [arXiv:2103.14386](http://arxiv.org/abs/2103.14386).
