### 2021-10-12 SNEWS2/snewpy#109 @apizzuto
Problem downloading models
Hi all, this is part of https://github.com/openjournals/joss-reviews/issues/3772. I was trying to install the code, and ran into some problems obtaining the models. I'm not sure if this is a bug or permissions issue, or if I misunderstood the install directions in the README, but here is a MWE reproducing the error:

```
MBP-C02F714BQ05Q:snewpy apizzuto$ python3 -m venv snewpy-venv
MBP-C02F714BQ05Q:snewpy apizzuto$ source snewpy-venv/bin/activate
(snewpy-venv) MBP-C02F714BQ05Q:snewpy apizzuto$ pip install snewpy
 . . .
(snewpy-venv) MBP-C02F714BQ05Q:snewpy apizzuto$ python -c 'import snewpy; snewpy.get_models()'
Available models in this version of SNEWPY: ['Bollig_2016', 'Fornax_2019', 'Fornax_2021', 'Kuroda_2020', 'Nakazato_2013', 'OConnor_2013', 'OConnor_2015', 'PISN', 'Sukhbold_2015', 'Tamborra_2014', 'Type_Ia', 'Walk_2018', 'Walk_2019', 'Warren_2020', 'Zha_2021']

Type a model name, 'all' to download all models or <Enter> to cancel: all
. . .
```
and then I get a bunch of messages along the lines of `Failed to download https://github.com/SNEWS2/snewpy/raw/v1.1b3/models/{model_name}/README.md to 'SNEWPY_models/{model_name}/{file_name}'.` After those outputs I get ```ERROR: 719 exceptions occured. ([URLError(SSLCertVerificationError(1, '[SSL: CERTIFICATE_VERIFY_FAILED] certificate verify failed: unable to get local issuer certificate (_ssl.c:1129)')) . . . ```
(I trimmed the entire message because it lists out all 719 errors). I have a subdirectory called `SNEWPY_models` now with all of the model names as subdirectories, but those subdirectories seem to be empty
