### 2021-10-12 3c194dc
JOSS paper improvements In response to https://github.com/openjournals/joss-reviews/issues/3772#issuecomment-940387588
