+++ doc/paper.bib
@@ -54,7 +54,7 @@
 }
 
 @misc{SNOwGLoBES,
-	author = "Scholberg, Kate and others",
+	author = "Scholberg, Kate and SNOwGLoBES Contributors",
 	journal = "GitHub repository",
 	publisher = "GitHub",
 	title = "{SNOwGLoBES code: SuperNova Observatories with GLoBES}",
