### 2021-10-12 SNEWS2/snewpy#109 @apizzuto
It looks like a way around this that doesn't introduce any dependencies would be adding `import ssl` with the other imports and then in the `except` block in https://github.com/SNEWS2/snewpy/blob/3e378d397fddad658775d8a9e39bd514d712915c/python/snewpy/__init__.py#L66 adding a
```
if (not os.environ.get('PYTHONHTTPSVERIFY', '') and getattr(ssl, '_create_unverified_context', None)):
    ssl._create_default_https_context = ssl._create_unverified_context
```

Happy to submit a pull request if the developers are happy with this patch
