### 2021-10-13 SNEWS2/snewpy#109 @apizzuto
Hi Jost, thanks for looking into this! I'm happy to close this issue because after your comment I think that it's my machine (OS: MacOS 11.2.3, python3.9.5). I recently started working on an M1 mac and it seems like issues like this always seem to crop up.
I tried the install on my previous machine (OS: macOS 10.15.3 and python3.8.5) and everything worked completely fine. I'll finish up the review on that machine or I'll add the context as an argument on my M1 and continue the review there. Thanks for being so prompt with this response!
