### 2021-10-13 #3772 @apizzuto
Hi @JostMigenda, thanks for the tweaks to the paper and looking into the install issue. My installation is working now and I checked off everything for the paper in my checklist. I'll work through the examples in the next few days.
