### 2021-10-13 SNEWS2/snewpy#111 @apizzuto
pip installation from PyPI leads to failed tests
Hello SNEWPY team! This is part of the JOSS review at https://github.com/openjournals/joss-reviews/issues/3772.

Using the suggested installation from the README (`pip install snewpy` and then `python -c 'import snewpy; snewpy.get_models()'` to grab models), the tests fail. Specifically, `test_models.py` outputs

```
FileNotFoundError: [Errno 2] No such file or directory: 'models/Nakazato_2013/nakazato-shen-z0.004-t_rev100ms-s13.0.fits'
```

I think this is just because when installing as directed in the README, the package distributed by PyPI doesn't have all of the included files in the github repo, and then when downloading the files the models go to a directory called `SNEWPY_models` instead of where the test script is looking (`models/`).

I think this is all okay, because I understand wanting to have only a limited version of the code available on pypi so that you don't need all of the large files from the examples / documentation / models. However, there is no alternative method described in the docs on how to install the code that would allow the user to run examples and run tests. Could you include a brief set of installation instructions in the README for those that would want access to the examples and such? I'm thinking something along the lines of what you do in the GitHub actions to automate tests, eg
```
git clone {SNEWPY_git_url}
python setup.py install --user
```
