### 2021-10-13 9c1341d
Improve README and documentation * Better introduction of README file * add installation instructions for developers * collapse instructions for installing (SNOw)GLoBES in README * add links to arXiv paper and readthedocs page
