### 2021-10-13 SNEWS2/snewpy#111 @JostMigenda
Hi @apizzuto, thanks for the suggestion! I’ve added the suggested instructions to the documentation and have made a few more updates to the README file while I was at it. Currently [on a separate branch](https://github.com/SNEWS2/snewpy/tree/JostMigenda/JOSSReview); I’ll leave that PR open for a while in case there are further opportunities to improve documentation.
