### 2021-10-13 ffa9a09
update Nakazato references in .ipynb and docstring
