### 2021-10-15 SNEWS2/snewpy#113 @apizzuto
Notebook documentation cleanup and minor notebook errors
Hey SNEWPY team! This is part of https://github.com/openjournals/joss-reviews/issues/3772.

These are a summary of some of the minor issues I ran into when working through the (truly impressive) notebooks that you all put together. First, I have a request for documentation: could you add a brief README to the directory where the notebooks live, briefly explaining what each one is for? It was a bit daunting at first when I saw there were so many and wasn't sure where to start. It's up to you, but it might be worth separating those that demonstrate core functionality (eg. `SNEWPY_examples.ipynb`) from those just highlighting a single model by putting all of the model ones in a subdirectory.

As for specific issues:
* `SNEWPY_examples.ipynb`: Would you be able to add a little bit of text to the beginning of the notebook (or throughout) that describes what's happening in this notebook? The inline comments are helpful, but some overview at the beginning would help provide context
* `snowglobes_models.ipynb`: all of the models with "Type Ia" had `FileNotFoundErrors` because it was a space instead of an underscore (sn5, sn6) in the filename
* `SNEWS2.0_whitepaper_rates`: I think you need a `from snewpy.models import Bollig2016`, otherwise in the Generate Fluence Files section I get a `NameError: name 'Bollig2016' is not defined`
* `FlavorTransformations.ipynb`: I'm guessing this is an artifact of a weird PYTHONPATH error from whoever wrote this notebook, but it looks like the imports don't work here because they are trying to import from `python.snewpy` instead of `snewpy`
* `Flavor_XForm` and all of the model notebooks: No issue here, I just really enjoyed working through them :)
* `Fornax_2019.ipynb`: `ImportError` because some of the later plots require `healpy`. Is this functionality crucial for SNEWPY and should `healpy` be a dependency? If not, can you just add a disclaimer to the top of the notebook about how this notebook will only work if the user has `healpy`?
* `Sukhbold_2015.ipynb`: `fig = plot_luminosity(model)` gives an error: `AttributeError: 'Sukhbold_2015' object has no attribute 'fitsfile'` from the title line in the call to `ax.plot`. Is it supposed to be `model.filename`?
* This is related to adding a general README, but I couldn't immediately tell what the difference between `Flavor_XForm.ipynb` and `FlavorTransformations.ipynb` were
