### 2021-10-15 #3772 @apizzuto
I just had a look at all of the notebooks, and first and foremost, after running through them I have to say I'm really impressed with all of the notebooks and all of the examples that you provided. I had a good time running through all of the different model notebooks and seeing what SNEWPY can do.
However, I ran into a couple of small problems with some of the notebooks, and I had a request for the organization of the notebooks. These are all outlined in the issue I just opened (https://github.com/SNEWS2/snewpy/issues/113).

I think that I'm close to finishing up my review, but just to keep everyone in the loop – I still want to (1) read through the rest of the API docs to make sure everything looks good, (2) briefly take a glance at the code to make sure all of the functionality claims are verified (although after looking at the notebooks I think it's safe to say that they are) and (3) take a glance at the testing suite
