### 2021-10-15 SNEWS2/snewpy#113 @apizzuto
I just realized that I forgot to tag the PR about documentation https://github.com/SNEWS2/snewpy/pull/112

Another small note: in the `SNEWS2.0_whitepaper_rates` notebook, there are some hardcoded paths in the "Execute SNOwGLoBES" section
