### 2021-10-18 SNEWS2/snewpy#114 @apizzuto
Test coverage – snewpy.snowglobes
Hi all! This is part of the JOSS paper review: https://github.com/openjournals/joss-reviews/issues/3772
As part of checking over the tests, I looked into the code coverage that they represent (printed below):
```Name                                      Stmts   Miss  Cover
-------------------------------------------------------------
python/snewpy/__init__.py                    77     69    10%
python/snewpy/_git.py                        73     73     0%
python/snewpy/_model_urls.py                  1      1     0%
python/snewpy/_version.py                     1      0   100%
python/snewpy/flavor_transformation.py      309     15    95%
python/snewpy/from_snowglobes.py              6      6     0%
python/snewpy/models.py                     720    273    62%
python/snewpy/neutrino.py                    44      4    91%
python/snewpy/run_snowglobes.py               6      6     0%
python/snewpy/snowglobes.py                 688    688     0%
python/snewpy/test/__init__.py                1      1     0%
python/snewpy/test/snewpy_test_suite.py      11     11     0%
python/snewpy/test/test_init.py               3      1    67%
python/snewpy/test/test_models.py           159      0   100%
python/snewpy/test/test_neutrino.py          38      0   100%
python/snewpy/test/test_xforms.py           308      0   100%
python/snewpy/to_snowglobes.py               15     15     0%
-------------------------------------------------------------
TOTAL                                      2460   1163    53%
```

I think for the most part this is sufficient, but I was a little concerned that there are no tests for the snewpy.snowglobes part of the package, because that seems to be a fairly substantive part of the code. I understand that having GitHub Actions do the snowglobes installation would be very difficult, but is there any way to write tests for the other functionality of `snewpy.snowglobes` – generating input files for SNOwGLoBES or collating SNOwGLoBES outputs?

If you all think this is totally unnecessary, I could see us closing this issue, but if it's easy enough to write some of these tests I think it might be worth adding.
