### 2021-10-18 #3772 @JostMigenda
As you say, detectors are all implemented in SNOwGLoBES, so that’s not part of SNEWPY. [SNOwGLoBES’ documentation](https://github.com/SNOwGLoBES/snowglobes/tree/master/doc) has info on how to add a detector, though.

Many thanks for the thorough check of example notebooks and tests! I’ll respond in the respective issues.
