### 2021-10-18 SNEWS2/snewpy#114 @JostMigenda
That’s a good comment; thanks!

As is, I think those functions are very hard to test properly, since their main purpose is a side effect (writing/reading files)—so testing would require shipping dozens of sample in-/output files to compare results with expectations. Testing properly would probably require refactoring large parts of the code, to try and isolate the parts that write to/read from files. Given that this is code that should otherwise not change much (mainly if the SNOwGLoBES file format changes, which historically has been very rare), I think that’s probably not worthwhile.

While developing SNEWPY, I guess we’ve effectively treated the sample scripts/notebooks as “integration tests” to verify that the end results (after creating flux files, running SNOwGLoBES and collating its outputs) are as expected. To me, this looks like the better way forward: Install (SNOw)GLoBES as part of a testing workflow and run one of the sample scripts as an integration workflow. (And longer term, perhaps even see whether we can ship compiled binaries for these dependencies as part of SNEWPY, to simplify the installation process? Though that’s not going to be possible on the time scale of the JOSS review.)
