### 2021-10-18 SNEWS2/snewpy#114 @JostMigenda
@sybenzvi I think you’ve written most of the tests, so I’d love a second opinion from you on this!
