### 2021-10-19 SNEWS2/snewpy#114 @apizzuto
Hi @sybenzvi, thanks for the input!
I think that it's good to have _some_ way to check that the code is working properly after new commits, but I don't think that _has_ to be with just the automated tests. As @JostMigenda mentioned, though, some of the examples in `scripts/` and `doc/nb` do essentially check the end-to-end behavior with using `snewpy.snowglobes`, so I think that for the purpose of this review, it is enough to have your examples serve as a sort of de facto testing suite for `snewpy.snowglobes`. It is up to you whether you want to leave them as is or if you want to add a few extra lines of markdown cells in the example notebooks saying "so-and-so output should be roughly x" so that people can actually verify that things are working properly.

I'm happy to hear that you might want to add this long term, but for the purpose of this review I'm happy to close the issue with the idea that folks can check that `snewpy.snowglobes` is behaving properly with the examples. If you all are happy with this, please feel free to close this issue!
