### 2021-10-19 cb5d24e
Update integration.yml
fix syntax error
