+++ .github/workflows/integration.yml
@@ -19,7 +19,7 @@ jobs:
     strategy:
       # Add a list of python versions we want to use for testing.
       matrix:
-        python-version: '3.9'
+        python-version: ['3.9']
 
     # Steps represent a sequence of tasks that will be executed as part of the job
     steps:
