### 2021-10-19 265e90e
adding test with hardcoded GitHub paths
