+++ .github/workflows/integration.yml
@@ -10,18 +10,14 @@ on:
   pull_request:
     branches: [ main ]
 
-# A workflow run is made up of one or more jobs that can run sequentially or in parallel
 jobs:
-  # This workflow contains a single job called "build"
   build:
-    # The type of runner that the job will run on.
     runs-on: ubuntu-latest
     strategy:
       # Add a list of python versions we want to use for testing.
       matrix:
         python-version: ['3.9']
 
-    # Steps represent a sequence of tasks that will be executed as part of the job
     steps:
       - uses: actions/checkout@v2
       - name: Set up Python ${{ matrix.python-version }}
@@ -37,8 +33,6 @@ jobs:
           python setup.py install --user
       - name: Install SNOwGLoBES
         run: |
-          pwd
-          echo $GITHUB_WORKSPACE
           mkdir opt
           cd opt
           wget https://www.mpi-hd.mpg.de/personalhomes/globes/download/globes-3.2.17.tar.gz
@@ -47,10 +41,9 @@ jobs:
           ./configure --prefix=$GITHUB_WORKSPACE/opt/globes-3.2.17-install  --disable-binary
           make
           make install
-          cd $GITHUB_WORKSPACE/opt/globes-3.2.17-install
-          export GLB_DIR=${PWD}
-          cd ..
+          export GLB_DIR=${GITHUB_WORKSPACE}/opt/globes-3.2.17-install
           # Install SNOwGLoBES
+          cd ${GITHUB_WORKSPACE}/opt/
           git clone https://github.com/SNOwGLoBES/snowglobes.git
           cd snowglobes
           export SNOWGLOBES=${PWD}
@@ -59,5 +52,4 @@ jobs:
           make install
       - name: Run Integration Test
         run: |
-          pwd
-          python python/snewpy/scripts/SNEWS2.0_rate_table_singleexample.py
+          python -m unittest python/snewpy/test/snowglobes_integrationtest.py
