+++ python/snewpy/test/snowglobes_integrationtest.py
@@ -13,7 +13,7 @@ class TestSNOwGLoBES(unittest.TestCase):
         """
         # Hardcoded paths on GitHub Action runner machines
         SNOwGLoBES_path = "/home/runner/work/snewpy/snewpy/opt/snowglobes/"
-        SNEWPY_model_dir = "/home/runner/work/snewpy/models/"
+        SNEWPY_model_dir = "/home/runner/work/snewpy/snewpy/models/"
 
         distance = 10  # Supernova distance in kpc
         detector = "wc100kt30prct" #SNOwGLoBES detector for water Cerenkov
