### 2021-10-19 90a81cd
fix normalization
