### 2021-10-20 34b9949
try to run integration test on all OSs
