+++ .github/workflows/integration.yml
@@ -12,10 +12,11 @@ on:
 
 jobs:
   build:
-    runs-on: ubuntu-latest
+    runs-on: ${{ matrix.os }}
     strategy:
       # Add a list of python versions we want to use for testing.
       matrix:
+        os: [ubuntu-latest, macos-latest, windows-latest]
         python-version: ['3.9']
 
     steps:
