### 2021-10-20 9f914fc
remove Windows from OS list
