+++ .github/workflows/integration.yml
@@ -12,11 +12,10 @@ on:
 
 jobs:
   build:
-    runs-on: ${{ matrix.os }}
+    runs-on: ubuntu-latest
     strategy:
       # Add a list of python versions we want to use for testing.
       matrix:
-        os: [ubuntu-latest, macos-latest]
         python-version: ['3.9']
 
     steps:
