### 2021-10-20 fcce104
Adding Integration Test (#115)
Create integration.yml and add test script with hardcoded paths for GitHub runners
