### 2021-10-21 209cd73
reimplement OConnor_2013 model set
this was lost somewhere along the line, now its back with a Notebook too
