### 2021-10-21 e8d438c
fix docstring, remove unused function from notebook
