### 2021-10-21 e42885b
document AnalyticFluence notebook better
