### 2021-10-22 6636cf2
Remove AnalyticModel notebook, see #119
