### 2021-11-01 #3772 @arfon
Just an FYI that the DOI for the related AAS publication will be `10.3847/1538-4357/ac350f`
