### 2021-11-01 0d0ac96
add AAS DOI; remove article from journal name since the JOSS template already includes it
