### 2021-11-01 6e4a3a9
fix reading from SNOwGLoBES files previous code would cut off initial digit of event count, see https://github.com/SNEWS2/snewpy/issues/122
