### 2021-11-03 26efd27
Make uniform cycle over the detectors in 'simulate'
