+++ python/snewpy/snowglobes.py
@@ -41,6 +41,7 @@ from snewpy.neutrino import Flavor, MassHierarchy
 
 mpl.use('Agg')
 
+logger = logging.getLogger(__name__)
 
 def generate_time_series(model_path, model_type, transformation_type, d, output_filename=None, ntbins=30, deltat=None):
     """Generate time series files in SNOwGLoBES format.
@@ -674,83 +675,31 @@ def simulate(SNOwGLoBESdir, tarball_path, detector_input="all", verbose=False):
     #This is the place where you can comment out any detectors you don't want to run
     #This runs the entire module, for each detector configuration
 
+
+    det_materials = { 'icecube':'water',
+                      'wc100kt30prct':'water',
+                      'wc100kt15prct':'water',
+                      'hyperk30prct':'water',
+                      'km3net':'water',
+                      'ar40kt':'argon',
+                      'novaND':'nova_soup',
+                      'novaFD':'nova_soup',
+                      'scint20k':'scint',
+                      'halo1':'lead',
+                      'halo2':'lead'
+                      }
+
+    if detector_input == "all":
+        detector_input=list(det_materials.keys())
+    if isinstance(detector_input,str):
+        detector_input=[detector_input]
+
+    logger.info(f'detector_input={detector_input}')
+
     for IndividualFluxFile in FluxFileNameStems:
-        if (verbose):
-            print(IndividualFluxFile)
-        position = FluxFileNameStems.index(IndividualFluxFile) + 1
-        total = len(FluxFileNameStems)
-        percent_calc = (position/total)*100
-        percentage_done = str(round(percent_calc, 2))
-
-        if detector_input == "all":
-            if (verbose):
-                print("Running all detectors")
-            if format_globes_for_supernova(IndividualFluxFile, "water", "icecube", "weight") == "Complete":
-                if verbose:
-                    print("Finished {0}||icecube".format(IndividualFluxFile))
-            if format_globes_for_supernova(IndividualFluxFile, "water", "wc100kt30prct", "weight") == "Complete":
-                if verbose:
-                    print("Finished {0}||wc100kt30prct".format(IndividualFluxFile))
-            #if format_globes_for_supernova(IndividualFluxFile, "water", "wc100kt30prct_he", "weight") == "Complete":
-                # if verbose: print("Finished {0}||wc100kt30prct_he".format(IndividualFluxFile))
-            if format_globes_for_supernova(IndividualFluxFile, "water", "wc100kt15prct", "weight") == "Complete":
-                if verbose:
-                    print("Finished {0}||wc100kt15prct".format(IndividualFluxFile))
-            #if format_globes_for_supernova(IndividualFluxFile, "water", "hyperk30prct", "weight") == "Complete": #not working
-                # if verbose: print("Finished {0}||hyperk30prct".format(IndividualFluxFile))
-            if format_globes_for_supernova(IndividualFluxFile, "argon", "ar40kt", "weight") == "Complete":
-                if verbose:
-                    print("Finished {0}||ar40kt".format(IndividualFluxFile))
-            if format_globes_for_supernova(IndividualFluxFile, "lead", "halo1", "weight") == "Complete":
-                if verbose:
-                    print("Finished {0}||halo1".format(IndividualFluxFile))
-            if format_globes_for_supernova(IndividualFluxFile, "lead", "halo2", "weight") == "Complete":
-                if verbose:
-                    print("Finished {0}||halo2".format(IndividualFluxFile))
-            #if format_globes_for_supernova(IndividualFluxFile, "argon", "ar40kt_he", "weight") == "Complete":
-                # if verbose: print("Finished {0}||ar40kt_he".format(IndividualFluxFile))
-            if format_globes_for_supernova(IndividualFluxFile, "scint", "scint20kt", "weight") == "Complete":
-                if verbose:
[…]
