### 2021-11-04 2f0f3ab
Refactoring format_globes_for_supernova
