+++ python/snewpy/snowglobes.py
@@ -28,6 +28,9 @@ import os
 import re
 import tarfile
 import zipfile
+from pathlib import Path
+from collections import namedtuple
+
 from subprocess import call
 
 import matplotlib as mpl
@@ -331,267 +334,104 @@ def simulate(SNOwGLoBESdir, tarball_path, detector_input="all", verbose=False):
     verbose : bool
         Whether to generate verbose output, e.g. for debugging.
     """
+    
+    sng = Path(SNOwGLoBESdir)
+    #prepare output folder
+    output_dir = sng/'out'
+    output_dir.mkdir(exist_ok=True)
+    #extract flux files
+    flux_dir = sng/'flux'
 
-    #Extracts data from tarfile and sets up lists of paths and fluxfilenames for later use
+    with tarfile.open(tarball_path,'r') as f:
+        flux_files = f.getnames() 
+        f.extractall(path=flux_dir)
+     
+    FluxFileNameStems = [Path(f).stem for f in flux_files if f.endswith('.dat')]
+    #regexps to use
+    re_flux = re.compile('flux_file=.*\n')
+    re_tgt  = re.compile('target_mass=.*\n')
 
-    if tarfile.is_tarfile(tarball_path):  # extracts tarfile
-        if (verbose):
-            print("Valid Tar file")
-        tar = tarfile.open(tarball_path)
-        TarballFileNames = tar.getnames()
-        tar.extractall(path=SNOwGLoBESdir + "/fluxes")
-        tar.close()
-        if (verbose):
-            print("Tar file data extracted")
-    elif zipfile.is_zipfile(tarball_path):  # extracts zipfile
-        if (verbose):
-            print("Valid Zip file")
-        zip = zipfile.ZipFile(tarball_path)
-        TarballFileNames = zip.namelist()
-        for fileName in TarballFileNames:
-            if str(fileName)[0] == "." or str(fileName)[0] == "_":
-                continue
-            else:
-                zip.extract(fileName, path=SNOwGLoBESdir + "/fluxes")
-                if (verbose):
-                    print("zipfile extracted")
-        zip.close()
-        if (verbose):
-            print("Zip file data extracted")
-    else:
-        print("Invalid Tarfile")  # doesn't accept anything other than tar and zip files
+    def load_channels(fname):
+        """load a table of channels from file"""
+        table = np.loadtxt(fname, dtype=str)
+        Channel = namedtuple('Channel',['name','number','parity','flavor','weight'])
+        return [Channel(*c) for c in table]
+    
+    def load_target_masses(fname):
+        t = np.loadtxt(fname, dtype=[('name','U100'),('mass','f'),('factor','f')])
+        tgt_mass = t['mass']*t['factor']
+        return dict(zip(t['name'],tgt_mass))
 
-    #Creates a new out folder if one does not exist
+    tgt_masses = load_target_masses(sng/'detector_configurations.dat')
 
-    out_folder = "{0}/out".format(SNOwGLoBESdir)
-    try:
-        os.mkdir(out_folder)
-    except OSError as e:
-        if (verbose):
-            print("out folder already exists")
-    else:
-        if (verbose):
-            print("out folder made")
+    def cat(fname):
+        with open(fname,'r') as f:
+            return f.read()
 
-    #Creates directories for each flux file within the out folder
-
-    FluxFileNameStems = []
-    #extension = '.dat'
-    for IndividualFile in TarballFileNames:
-        extension_beginning = str(IndividualFile).rfind(".")
-        if extension_beginning > -1:
-            extension = str(IndividualFile[extension_beginning:])
-            ExtensionFound = str(IndividualFile).rfind(extension)
-            # gets rid of extension at the end of flux files
[…]
