### 2021-11-04 2647f21
Fixed typos and errors
