### 2021-11-04 839d14d
Refactored apply_weights method.
