+++ python/snewpy/snowglobes.py
@@ -352,10 +352,10 @@ def simulate(SNOwGLoBESdir, tarball_path, detector_input="all", verbose=False):
     re_tgt  = re.compile('target_mass\s?=.*\n')
 
     def load_channels(fname):
-        """load a table of channels from file"""
-        table = np.loadtxt(fname, dtype=str)
-        Channel = namedtuple('Channel',['name','number','parity','flavor','weight'])
-        return [Channel(*c) for c in table]
+        logger.info(f'reading channels from {fname}')
+        table = np.loadtxt(fname, dtype=[('name','U100'),('number',int),('parity','U1'),('flavor','U1'),('weight',float)])
+        logger.info(f'channels={table}')
+        return table
     
     def load_target_masses(fname):
         t = np.loadtxt(fname, dtype=[('name','U100'),('mass','f'),('factor','f')])
@@ -363,7 +363,6 @@ def simulate(SNOwGLoBESdir, tarball_path, detector_input="all", verbose=False):
         return dict(zip(t['name'],tgt_mass))
 
     tgt_masses = load_target_masses(sng/'detector_configurations.dat')
-    logger.info(f'Target masses are: {tgt_masses}')
 
     def cat(fname):
         with open(fname,'r') as f:
@@ -376,9 +375,13 @@ def simulate(SNOwGLoBESdir, tarball_path, detector_input="all", verbose=False):
     # this function runs supernova
     def format_globes_for_supernova(FluxNameStem, ChannelName, ExpConfigName, NoWeight=1, extension=extension):
 
-        def prepare_globes_file(output, flux_fname, channel_fname, detector_name):
+
+        channel_fname = sng/f'channels/channels_{ChannelName}.dat'
+        channels = load_channels(channel_fname)
+
+        def prepare_globes_file(output, flux_fname, channels, detector_name):
             """Function to write the snowglobes configuration to file """
-            logger.info(f'Prepare globes run for [flux={flux_fname}, chan={channel_fname},det={detector_name}]')
+            logger.info(f'Prepare globes run for [flux={flux_fname}, det={detector_name}]')
             if flux_fname.is_file()==False:
                 raise FileNotFoundError(flux_fname)
             #write preamble
@@ -388,16 +391,14 @@ def simulate(SNOwGLoBESdir, tarball_path, detector_input="all", verbose=False):
             s = re_flux.sub(f'flux_file=   "{flux_fname.relative_to(sng)}"\n',s)
             output.write(s)
 
-            channels = load_channels(channel_fname)
             # smearing for each channel
             for c in channels:
-                output.write(f'include "{sng}/smear/smear_{c.name}_{detector_name}.dat"\n')
+                output.write(f'include \"{sng}/smear/smear_{c["name"]}_{detector_name}.dat"\n')
 
             # Writes modified DETECTOR to GLOBESFILE; replace the mass with the calculated TargetMass
             s = cat(sng/'glb/detector.glb')
             s = re_tgt.sub(f'target_mass=  {tgt_masses[detector_name]:13.6f}\n',s)
             output.write(s)
-
 
             # Now the cross-sections.  Note that some of these are repeated even though
             # it is not necessary (xscns for several flavors can be in the same file).
@@ -405,15 +406,15 @@ def simulate(SNOwGLoBESdir, tarball_path, detector_input="all", verbose=False):
 
             output.write("\n\n /******** Cross-sections *********/\n \n")
             for c in channels:
-                output.write(f'cross(#{c.name})< @cross_file= "{sng}/xscns/xs_{c.name}.dat" >\n')
+                output.write(f'cross(#{c["name"]})< @cross_file= "{sng}/xscns/xs_{c["name"]}.dat" >\n')
[…]
