+++ python/snewpy/snowglobes.py
@@ -29,7 +29,8 @@ import re
 import tarfile
 import zipfile
 from pathlib import Path
-from collections import namedtuple
+import subprocess
+import traceback
 
 from subprocess import call
 
@@ -339,22 +340,20 @@ def simulate(SNOwGLoBESdir, tarball_path, detector_input="all", verbose=False):
     #prepare output folder
     output_dir = sng/'out'
     output_dir.mkdir(exist_ok=True)
-    #extract flux files
     flux_dir = sng/'flux'
-
+    #extract flux files
     with tarfile.open(tarball_path,'r') as f:
         flux_files = f.getnames() 
         f.extractall(path=flux_dir)
      
-    FluxFileNameStems = [Path(f).stem for f in flux_files if f.endswith('.dat')]
+    flux_files = [Path(f) for f in flux_files if f.endswith('.dat')]
+    logger.debug(f'Flux files = {flux_files}')
     #regexps to use
     re_flux = re.compile('flux_file\s?=.*\n')
     re_tgt  = re.compile('target_mass\s?=.*\n')
 
     def load_channels(fname):
-        logger.info(f'reading channels from {fname}')
         table = np.loadtxt(fname, dtype=[('name','U100'),('number',int),('parity','U1'),('flavor','U1'),('weight',float)])
-        logger.info(f'channels={table}')
         return table
     
     def load_target_masses(fname):
@@ -364,84 +363,53 @@ def simulate(SNOwGLoBESdir, tarball_path, detector_input="all", verbose=False):
 
     tgt_masses = load_target_masses(sng/'detector_configurations.dat')
 
-    def cat(fname):
-        with open(fname,'r') as f:
-            return f.read()
+    def prepare_globes_file(output, flux_fname, channels, detector_name):
+        """Function to write the snowglobes configuration to file """
+        logger.debug(f'Prepare globes run for [flux={flux_fname}, det={detector_name}]')
 
-    extension='dat'
-    ##################################################################################################
-    #The main function that reformats data from a bunch of different files in order to run supernova.c
+        def cat(fname):
+            with open(fname,'r') as f:
+                return f.read()
+        if flux_fname.is_file()==False:
+            raise FileNotFoundError(flux_fname)
+        #write preamble
+        output.write(cat(sng/'glb/preamble.glb'))
+        #write flux, substituted with input argument
+        s = cat(sng/'glb/flux.glb')
+        s = re_flux.sub(f'flux_file=   "{flux_fname.relative_to(sng)}"\n',s)
+        output.write(s)
 
-    # this function runs supernova
-    def format_globes_for_supernova(FluxNameStem, ChannelName, ExpConfigName, NoWeight=1, extension=extension):
+        # smearing for each channel
+        for c in channels:
+            output.write(f'include \"{sng}/smear/smear_{c["name"]}_{detector_name}.dat"\n')
 
+        # Writes modified DETECTOR to GLOBESFILE; replace the mass with the calculated TargetMass
+        s = cat(sng/'glb/detector.glb')
+        s = re_tgt.sub(f'target_mass=  {tgt_masses[detector_name]:13.6f}\n',s)
+        output.write(s)
 
-        channel_fname = sng/f'channels/channels_{ChannelName}.dat'
-        channels = load_channels(channel_fname)
+        # Now the cross-sections.  Note that some of these are repeated even though
+        # it is not necessary (xscns for several flavors can be in the same file).
+        # This is just to make a consistent loop over channels.
+        output.write("\n\n /******** Cross-sections *********/\n \n")
+        for c in channels:
+            output.write(f'cross(#{c["name"]})< @cross_file= "{sng}/xscns/xs_{c["name"]}.dat" >\n')
 
-        def prepare_globes_file(output, flux_fname, channels, detector_name):
[…]
