### 2021-11-04 7d605cd
Refactoring collate() using numpy to read tables fast
