+++ python/snewpy/snowglobes.py
@@ -31,8 +31,7 @@ import zipfile
 from pathlib import Path
 import subprocess
 import traceback
-
-from subprocess import call
+import pdb
 
 import matplotlib as mpl
 import matplotlib.pyplot as plt
@@ -489,41 +488,15 @@ def collate(SNOwGLoBESdir, tarball_path, detector_input="all", skip_plots=False,
     dict
         Dictionary of data tables: One table per time bin; each table contains in the first column the energy bins, in the remaining columns the number of events for each interaction channel in the detector.
     """
-    model_dir, tarball = os.path.split(os.path.abspath(tarball_path))
-
+    sng = Path(SNOwGLoBESdir)
     #Determines type of input file
 
-    if ".tar.bz2" in str(tarball):
-        outputnamestem = tarball[0:str(tarball).rfind(".tar.bz2")]
-        tar = tarfile.open(tarball_path)
-        TarballFileNames = tar.getnames()
-        tar.close()
-    elif ".tar.gz" in str(tarball):
-        outputnamestem = tarball[0:str(tarball).rfind(".tar.gz")]
-        tar = tarfile.open(tarball_path)
-        TarballFileNames = tar.getnames()
-        tar.close()
-    elif ".zip" in str(tarball):
-        outputnamestem = tarball[0:str(tarball).rfind(".zip")]
-        zip = zipfile.ZipFile(tarball_path)
-        TarballFileNames = zip.namelist()
-        zip.close()
-    else:
-        print("Invalid Tar file")
+    with tarfile.open(tarball_path,'r') as f:
+        flux_files = f.getnames() 
 
+    flux_files = [f for f in flux_files if f.endswith('.dat')]
     #Determining the configuration of the files and directories within the tarfile and defining variables for uniformity
-
-    FluxFileNameStems = []
-    extension = '.dat'
-    for IndividualFile in TarballFileNames:
-        extension_beginning = str(IndividualFile).rfind(".")
-        if extension_beginning > -1:
-            extension = str(IndividualFile[extension_beginning:])
-            ExtensionFound = str(IndividualFile).rfind(extension)
-            # gets rid of extension at the end of flux files
-            FluxFileNameStems.append(str(IndividualFile)[0:ExtensionFound])
-        else:
-            continue
+    FluxFileNameStems = [Path(f).stem for f in flux_files]
 
     smearvals = ["_smeared_w", "_smeared_u", "nts_w", "nts_u"]
     #weightvals = ["_weighted", "unweighted"]
@@ -532,7 +505,6 @@ def collate(SNOwGLoBESdir, tarball_path, detector_input="all", skip_plots=False,
 
     #Add_funct sums up relevant files from output generated above
     def add_funct(flux, detector, smear, *arg):
-        homebase = SNOwGLoBESdir + "/out/"
 
         #Defining different variables for the combinatoric iterations
         #And reformatting some variables for different naming uses
@@ -571,91 +543,32 @@ def collate(SNOwGLoBESdir, tarball_path, detector_input="all", skip_plots=False,
         colors = ["k", "r", "g", "y", "b", "m", "c"]  # colors of graphed values
         compile_dict = {}
 
-        #Splits values from each file into Energy and Events, then sums appropriate ones
+        sng = Path(SNOwGLoBESdir)
+        homebase = sng/'out'
+        events_all = []
         for input_val in arg:
-            flux_filenames = os.listdir(homebase)
-            final_dict = {}
-            temp_dict = {}
-            for afile in flux_filenames:  # Does the actual summing of essential values in each file
-                name_iteration = "{0}_{1}*{2}*{3}*".format(flux, input_val, detector, smear)
-                if fnmatch.fnmatch(afile, name_iteration):  # and "Collated" not in str(afile):
[…]
