### 2021-11-04 9c0d7a4
Refactoring collate() - files handling
