+++ python/snewpy/snowglobes.py
@@ -31,6 +31,7 @@ import zipfile
 from pathlib import Path
 import subprocess
 import traceback
+import itertools
 import pdb
 
 import matplotlib as mpl
@@ -319,6 +320,25 @@ def generate_fluence(model_path, model_type, transformation_type, d, output_file
 
     return os.path.join(model_dir, tfname)
 
+det_materials = { 'icecube':'water',
+                  'wc100kt30prct':'water',
+                  'wc100kt15prct':'water',
+                  'hyperk30prct':'water',
+                  'km3net':'water',
+                  'ar40kt':'argon',
+                  'novaND':'nova_soup',
+                  'novaFD':'nova_soup',
+                  'scint20kt':'scint',
+                  'halo1':'lead',
+                  'halo2':'lead'
+                  }
+
+categories_map = {'water': ["nc_", "*_e_", "ibd", "nue_O16", "nuebar_O16"],
+                  'lead' : ["nc_", "*_e_", "nue_Pb208_1n", "nue_Pb208_2n"],
+                  'argon': ["nc_", "*_e_", "nue_Ar40", "nuebar_Ar40"],
+                  'scint': ["nc_", "*_e_", "ibd", "nue_C12", "nuebar_C12", "nue_C13"],
+                  'nova_soup':["nc_", "*_e_", "ibd", "nue_C12", "nuebar_C12"]
+                 }
 
 def simulate(SNOwGLoBESdir, tarball_path, detector_input="all", verbose=False):
     """Takes as input the neutrino flux files and configures and runs the supernova script inside SNOwGLoBES, which outputs calculated event rates expected for a given (set of) detector(s). These event rates are given as a function of the neutrino energy and time, for each interaction channel.
@@ -339,7 +359,7 @@ def simulate(SNOwGLoBESdir, tarball_path, detector_input="all", verbose=False):
     #prepare output folder
     output_dir = sng/'out'
     output_dir.mkdir(exist_ok=True)
-    flux_dir = sng/'flux'
+    flux_dir = sng/'fluxes'
     #extract flux files
     with tarfile.open(tarball_path,'r') as f:
         flux_files = f.getnames() 
@@ -421,9 +441,10 @@ def simulate(SNOwGLoBESdir, tarball_path, detector_input="all", verbose=False):
                 events[:,1]*=c["weight"]
                 np.savetxt(output_fname, events, fmt='%11g')
 
-    def run_simulation(flux_fname, channel_name, detector_name, do_apply_weight=False):
+    def run_simulation(flux_fname, detector_name, do_apply_weight=False):
         """this function runs supernova"""
-        channel_fname = sng/f'channels/channels_{channel_name}.dat'
+        material = det_materials[detector_name]
+        channel_fname = sng/f'channels/channels_{material}.dat'
         channels = load_channels(channel_fname)
         globes_fname = sng/'supernova.glb'
         with open(globes_fname,'w') as output:
@@ -440,19 +461,7 @@ def simulate(SNOwGLoBESdir, tarball_path, detector_input="all", verbose=False):
     #THE DRIVER
     #This is the place where you can comment out any detectors you don't want to run
     #This runs the entire module, for each detector configuration
-    det_materials = { 'icecube':'water',
-                      'wc100kt30prct':'water',
-                      'wc100kt15prct':'water',
-                      'hyperk30prct':'water',
-                      'km3net':'water',
-                      'ar40kt':'argon',
-                      'novaND':'nova_soup',
-                      'novaFD':'nova_soup',
-                      'scint20kt':'scint',
-                      'halo1':'lead',
-                      'halo2':'lead'
-                      }
-
+  
     if detector_input == "all":
[…]
