### 2021-11-04 41ffb2d
Use more floating point precision. Remove tracing
