### 2021-11-04 742bc95
Moved all loading functions to global scope
