### 2021-11-05 dcd9c62
Rearranged functions to keep all drawing stuff in one place
