+++ python/snewpy/snowglobes.py
@@ -333,11 +333,11 @@ det_materials = { 'icecube':'water',
                   'halo2':'lead'
                   }
 
-categories_map = {'water': ["nc_", "*_e_", "ibd", "nue_O16", "nuebar_O16"],
-                  'lead' : ["nc_", "*_e_", "nue_Pb208_1n", "nue_Pb208_2n"],
-                  'argon': ["nc_", "*_e_", "nue_Ar40", "nuebar_Ar40"],
-                  'scint': ["nc_", "*_e_", "ibd", "nue_C12", "nuebar_C12", "nue_C13"],
-                  'nova_soup':["nc_", "*_e_", "ibd", "nue_C12", "nuebar_C12"]
+categories_map = {'water': ["nc_*", "*_e", "ibd", "nue_O16", "nuebar_O16"],
+                  'lead' : ["nc_*", "*_e", "nue_Pb208_1n", "nue_Pb208_2n"],
+                  'argon': ["nc_*", "*_e", "nue_Ar40", "nuebar_Ar40"],
+                  'scint': ["nc_*", "*_e", "ibd", "nue_C12", "nuebar_C12", "nue_C13"],
+                  'nova_soup':["nc_*", "*_e", "ibd", "nue_C12", "nuebar_C12"]
                  }
 
 def load_datafile(f):
@@ -353,6 +353,17 @@ def load_target_masses(fname):
     tgt_mass = t['mass']*t['factor']
     return dict(zip(t['name'],tgt_mass))
 
+
+def _gen_datafile_name(flux,channel=None,detector=None, is_smeared:bool=False,is_weighted:bool=False):
+    """generate the name of snowglobes output data file"""
+    #list the path components
+    tokens = [flux,channel,detector,
+              'events_smeared' if is_smeared else 'events',
+              'weighted' if is_weighted else 'unweighted']
+    #remove empty tokens
+    tokens = [t for t in tokens if t is not '']
+    #assemble filename
+    return '_'.join(tokens)+'.dat'
 
 def simulate(SNOwGLoBESdir, tarball_path, detector_input="all", verbose=False):
     """Takes as input the neutrino flux files and configures and runs the supernova script inside SNOwGLoBES, which outputs calculated event rates expected for a given (set of) detector(s). These event rates are given as a function of the neutrino energy and time, for each interaction channel.
@@ -438,9 +449,7 @@ def simulate(SNOwGLoBESdir, tarball_path, detector_input="all", verbose=False):
 
     def apply_weights(detector_name, flux_fname, channels, do_smear):
         for c in channels:
-            input_fname = sng/'out/{fl}_{ch}_{det}_events{sm}_unweighted.dat'.format(
-                                        fl=flux_fname.stem,ch=c["name"], det=detector_name,
-                                        sm="_smeared" if do_smear else "")
+            input_fname = sng/'out'/_gen_datafile_name(flux_fname.stem,c["name"],detector_name,do_smear,is_weighted=False)
             output_fname = str(input_fname).replace('unweighted','weighted')
             data = load_datafile(input_fname)
             if len(data):
@@ -510,44 +519,22 @@ def collate(SNOwGLoBESdir, tarball_path, detector_input="all", skip_plots=False,
         flux_files = f.getnames() 
     flux_files = [f for f in flux_files if f.endswith('.dat')]
 
-
     FilesToCleanup = []
 
-    def add_funct(flux, detector, smear, *arg):
-        """Add_funct sums up relevant files from output generated above"""
-        #Defining different variables for the combinatoric iterations
-        #And reformatting some variables for different naming uses
-        if smear.endswith('_w'):
-            weight_value = "weighted"
-        else:
-            weight_value = "unweighted"
[…]
