### 2021-11-05 411d891
Enable cleanup, remove pdb.set_trace
