### 2021-11-05 6c82ab8
Catch exceptions during
