### 2021-11-05 dafbf80
Fix div by zero in Warren_2020 models.
