### 2021-11-06 5d3e78a
integration test with comparison of output archive to ref
