+++ python/snewpy/test/snowglobes_integration_benchmark_test.py
@@ -11,42 +11,27 @@ outdir = Path(snowglobes_dir)/'out'
 
 #simulation setup
 model_file = basedir/'models/Kuroda_2020/LnuR10B12.dat'
-model_type =model_file.parent.stem
+model_type = model_file.parent.stem
 xform_type = 'AdiabaticMSW_NMO'
+detectors= 'all'
 
 fluence_name = f'test_{model_type}_{xform_type}'
 fluence_file = model_file.parent/f'{fluence_name}.tar.bz2'
-detectors= 'all'
+output_archive = model_file.parent/f'{fluence_name}_SNOprocessed.tar.gz'
 
 #checks setup
-ref_dir = basedir/'out_ref1' 
+reference_archive = basedir/f'reference_{fluence_name}.tar.gz' 
 
 from snewpy import snowglobes
 
-#main steps defined here
-def generate():
-    f = snowglobes.generate_fluence(model_file, model_type, xform_type, d=10, 
-                                    output_filename=fluence_name)
-    return f
-
-def simulate():
-    #cleanup output directory first
-    for f in outdir.glob('*.*'):
-        f.unlink()
-    snowglobes.simulate(snowglobes_dir,fluence_file, detector_input=detectors)
-
-def collate():
-    #recreate the file to avoid crashes
-    Path(f'{snowglobes_dir}/fluxes/{fluence_name}.dat').touch()
-    try:
-        result = snowglobes.collate(snowglobes_dir,fluence_file, detector_input=detectors, 
-                                            skip_plots=True, remove_generated_files=False)
-    finally:
-        #return back to original directory
-        os.chdir(basedir)
-    return result
-
-def read_data(fname):
+#-------------------------------------------------
+def _cleanup(path, pattern='*.*'):
+    "Remove all the files from path/ matching pattern"
+    if path.exists():
+        for f in path.glob(pattern):
+            f.unlink()
+    
+def _read_data(fname):
     return np.loadtxt(fname,comments=['---','Total','#','Energy'], dtype='f8')
 
 def _check_file_in_directory(ref_file, check_dir):
@@ -61,6 +46,28 @@ def _check_file_in_directory(ref_file, check_dir):
         d1 = read_data(check_file)
         print(d0-d1)
         assert np.allclose(d0,d1)
+#--------------------------------------------
+#main steps defined here
+def generate():
+    f = snowglobes.generate_fluence(model_file, model_type, xform_type, d=10, 
+                                    output_filename=fluence_name)
+    return f
+
+def simulate():
+    #cleanup output directory first
+    _cleanup(outdir)
+    snowglobes.simulate(snowglobes_dir,fluence_file, detector_input=detectors)
+
+def collate():
+    #recreate the file to avoid crashes
+    Path(f'{snowglobes_dir}/fluxes/{fluence_name}.dat').touch()
+    try:
+        result = snowglobes.collate(snowglobes_dir,fluence_file, detector_input=detectors, 
+                                            skip_plots=True, remove_generated_files=False)
+    finally:
+        #return back to original directory
+        os.chdir(basedir)
+    return result
 
 #-------------------------------------------
 @pytest.mark.timing
@@ -75,6 +82,7 @@ def test_benchmark_simulate(benchmark):
 def test_benchmark_collate(benchmark):
     benchmark(collate)
 #-------------------------------------------
+@pytest.mark.snoglobes_steps
 def test_generate():
     #cleanup fluence file
     fluence_file.unlink(missing_ok=True)
@@ -83,15 +91,61 @@ def test_generate():
     assert result == str(fluence_file)
     assert fluence_file.exists()
 
+@pytest.mark.snoglobes_steps
+def test_simulate():
+    simulate()
[…]
