+++ .github/workflows/integration.yml
@@ -27,6 +27,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
+          pip install pytest pytest-benchmark 
           if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
       - name: Install SNEWPY
         run: |
@@ -50,6 +51,12 @@ jobs:
           cd src
           make
           make install
-      - name: Run Integration Test
+      - name: Run Integration Test#1
         run: |
           python -m unittest python/snewpy/test/snowglobes_integrationtest.py
+      - name: Run Integration Test#2
+        run: pytest -m 'crosscheck' 
+      - name: Run benchmark test
+        run: pytest --benchmark-only --benchmark-json output.json
+      - name: Publish benchmark results
+        uses: nils-braun/pytest-benchmark-commenter@v2
