### 2021-11-06 7c31629
Skip tests if $SNOWGLOBES variable not set
