### 2021-11-06 b21d7c8
Remove skipping tests, it should fail if running without SNOGLOBES
