### 2021-11-06 3dfdadf
Put integration tests in a single step
