### 2021-11-06 80d87a6
Fixed integration test: looking in subfolders
