### 2021-11-07 e648353
Adding new interface for snowglobes
