### 2021-11-07 d126da8
Adding the template to package, so it can be read
