+++ requirements.txt
@@ -1,5 +1,6 @@
 numpy >= 1.16
 scipy
 astropy
+pandas == 1.3.4
 matplotlib
 h5py
