### 2021-11-07 9e19c45
Fix errors when parsing filenames. More logging.
