+++ python/snewpy/snowglobes_interface.py
@@ -28,7 +28,8 @@ class SNOwGLoBES:
         df = pd.read_table(path,names=['name','mass','factor'], delim_whitespace=True, comment='#')
         df['tgt_mass']=df.mass*df.factor
         self.detectors=df.set_index('name').T
-        logger.debug(f'read detectors: {self.detectors}')
+        logger.info(f'read masses for detectors {self.detectors}')
+        logger.debug(f'detectors: {self.detectors}')
        
     def _load_channels(self, chan_dir):
         self.channels = {}
@@ -39,21 +40,29 @@ class SNOwGLoBES:
             self.channels[material] = df
         self.materials = list(self.channels.keys())
         self.chan_dir = chan_dir
-        logger.info(f'read materials: {self.materials}')
-        logger.debug(f'read channels: {self.channels}')
+        logger.info(f'read channels for  materials: {self.materials}')
+        logger.debug(f'channels: {self.channels}')
         
     def _load_efficiencies(self, path):
         result = {}
-        for file in path.glob(f'effic_*_*.dat'):
-            channel,detector = file.stem[len('effic_'):].rsplit('_',1)
-            res_det = result.get(detector,{})
-            with open(file) as f:
-                eff_string = f.read()
-                res_det[channel]= eff_string
+        for detector in self.detectors:
+            res_det = {}
+            for file in path.glob(f'effic_*_{detector}.dat'):
+                channel =  file.stem[len('effic_'):-len(detector)-1]
+                logger.debug(f'Reading file ({detector},{channel}): {file}')
+                with open(file) as f:
+                    res_det[channel]= f.read()
             result[detector]=res_det
         self.efficiencies = result 
+        logger.info(f'read efficiencies for materials: {list(self.efficiencies.keys())}')
+        logger.debug(f'efficiencies: {self.efficiencies}')
        
     def run(self, flux_file, detector, material):
+        assert material in self.materials
+        assert detector in self.detectors
+        assert detector in self.efficiencies
+
+        assert Path(flux_file).exists()
         return Runner(self,Path(flux_file),detector,material).run()
 
 @dataclass    
@@ -69,12 +78,8 @@ class Runner:
         self.det_config=self.sng.detectors[self.detector]
         self.base_dir=self.sng.base_dir
         self.out_dir=self.base_dir/'out'
-        self.flux_file = self.flux_file.resolve()
-        if self.flux_file.exists()==False:
-            raise FileNotFoundError(self.flux_file)
-
     def _generate_globes_config(self):
-        cfg =  self.sng.template.render(flux_file=self.flux_file.relative_to(self.base_dir),
+        cfg =  self.sng.template.render(flux_file=self.flux_file.resolve(),
                                     detector=self.detector,
                                     target_mass=self.det_config.tgt_mass,
                                     smear_dir=self.base_dir/'smear',
@@ -84,17 +89,22 @@ class Runner:
         return cfg
 
     def _parse_output(self, output):
-        #read the generated files
         data = {}
         for l in output:
+            #read the generated files from output
             if l.endswith('weighted.dat'):
                 channum, fname = l.split()
-                E,N = np.loadtxt(self.base_dir/fname, comments=['--','Total'], unpack=True)
+                fname = self.base_dir/fname
+                #load the data from file
+                try:
+                    E,N = np.loadtxt(fname, comments=['--','Total'], unpack=True)
+                finally:
+                    fname.unlink() #cleanup file
                 channel=self.channels.loc[int(channum)]
-                smeared= 'smeared'  if '_smeared' in fname else 'unsmeared'
+                smeared= 'smeared'  if '_smeared' in fname.stem else 'unsmeared'
[…]
