### 2021-11-07 a2a6de0
Adding an integration test: check the total number of events
