### 2021-11-07 e28cafd
Additional logging and handling errors
