### 2021-11-08 672c2c6
Tests fixture for snowglobes object
