+++ python/snewpy/snowglobes.py
@@ -41,6 +41,7 @@ from snewpy.neutrino import Flavor, MassHierarchy
 
 mpl.use('Agg')
 
+logger = logging.getLogger(__name__)
 
 def generate_time_series(model_path, model_type, transformation_type, d, output_filename=None, ntbins=30, deltat=None):
     """Generate time series files in SNOwGLoBES format.
@@ -316,6 +317,26 @@ def generate_fluence(model_path, model_type, transformation_type, d, output_file
     return os.path.join(model_dir, tfname)
 
 
+from snewpy.snowglobes_interface import SNOwGLoBES
+from pathlib import Path
+from tqdm import tqdm
+from tempfile import TemporaryDirectory
+
+def get_material(detector):
+    if detector.startswith('wc') or detector.startswith('ice'):
+        return 'water'
+    elif detector.startswith('d2O'):
+        return 'heavywater'
+    elif detector.startswith('ar'):
+        return 'argon'
+    elif detector.startswith('nova'):
+        return 'nova_soup'
+    elif detector.startswith('halo'):
+        return 'lead'
+    elif detector.startswith('scint'):
+        return 'scint'
+
+
 def simulate(SNOwGLoBESdir, tarball_path, detector_input="all", verbose=False):
     """Takes as input the neutrino flux files and configures and runs the supernova script inside SNOwGLoBES, which outputs calculated event rates expected for a given (set of) detector(s). These event rates are given as a function of the neutrino energy and time, for each interaction channel.
 
@@ -330,426 +351,27 @@ def simulate(SNOwGLoBESdir, tarball_path, detector_input="all", verbose=False):
     verbose : bool
         Whether to generate verbose output, e.g. for debugging.
     """
-
+    
+    sng = SNOwGLoBES(SNOwGLoBESdir)
+    if detector_input == 'all':
+        detector_input = list(sng.detectors)
+    elif isinstance(detector_input,str):
+        detector_input = [detector_input]
+    
     #Extracts data from tarfile and sets up lists of paths and fluxfilenames for later use
-
-    if tarfile.is_tarfile(tarball_path):  # extracts tarfile
-        if (verbose):
-            print("Valid Tar file")
-        tar = tarfile.open(tarball_path)
-        TarballFileNames = tar.getnames()
-        tar.extractall(path=SNOwGLoBESdir + "/fluxes")
-        tar.close()
-        if (verbose):
-            print("Tar file data extracted")
-    elif zipfile.is_zipfile(tarball_path):  # extracts zipfile
-        if (verbose):
-            print("Valid Zip file")
-        zip = zipfile.ZipFile(tarball_path)
-        TarballFileNames = zip.namelist()
-        for fileName in TarballFileNames:
-            if str(fileName)[0] == "." or str(fileName)[0] == "_":
-                continue
-            else:
-                zip.extract(fileName, path=SNOwGLoBESdir + "/fluxes")
-                if (verbose):
-                    print("zipfile extracted")
-        zip.close()
-        if (verbose):
-            print("Zip file data extracted")
-    else:
-        print("Invalid Tarfile")  # doesn't accept anything other than tar and zip files
-
-    #Creates a new out folder if one does not exist
-
-    out_folder = "{0}/out".format(SNOwGLoBESdir)
-    try:
-        os.mkdir(out_folder)
-    except OSError as e:
-        if (verbose):
-            print("out folder already exists")
-    else:
-        if (verbose):
-            print("out folder made")
-
-    #Creates directories for each flux file within the out folder
-
-    FluxFileNameStems = []
-    #extension = '.dat'
-    for IndividualFile in TarballFileNames:
-        extension_beginning = str(IndividualFile).rfind(".")
[…]
