### 2021-11-08 3153082
Raise exception if run failed
