### 2021-11-08 62a51f1
Collate function now creates tar files, as before
