+++ python/snewpy/snowglobes.py
@@ -368,7 +368,7 @@ def simulate(SNOwGLoBESdir, tarball_path, detector_input="all", verbose=False):
         for det in tqdm(detector_input, desc='Detectors'):
             mat = get_material(det)
             res=sng.run(flux_files, det, mat)
-            result[det]=dict(zip(flux_files,res))
+            result[det]=dict(zip((f.name for f in flux_files),res))
     return result 
 
 
@@ -395,331 +395,50 @@ def collate(SNOwGLoBESdir, tarball_path, detector_input="all", skip_plots=False,
     dict
         Dictionary of data tables: One table per time bin; each table contains in the first column the energy bins, in the remaining columns the number of events for each interaction channel in the detector.
     """
-    model_dir, tarball = os.path.split(os.path.abspath(tarball_path))
 
-    #Determines type of input file
+    def aggregate_channels(table, **patterns):
+        #rearrange the table to have only channel column
+        levels = list(table.columns.names)
+        levels.remove('channel')
+        t = table.stack(levels)
+        for name,pattern in patterns.items():
+            #get channels which contain `like`
+            t_sel = t.filter(like=pattern)
+            #sum over them and save to a separate column
+            t_agg = t_sel.sum(axis='columns')
+            #drop processed channels
+            t.drop(t_sel.columns, axis='columns',inplace=True)
+            t[name]=t_agg #fill the column
+        #return table with the original levels order
+        t = t.unstack(levels)
+        t = t.reorder_levels(table.columns.names, axis=1)
+        return t
 
-    if ".tar.bz2" in str(tarball):
-        outputnamestem = tarball[0:str(tarball).rfind(".tar.bz2")]
-        tar = tarfile.open(tarball_path)
-        TarballFileNames = tar.getnames()
-        tar.close()
-    elif ".tar.gz" in str(tarball):
-        outputnamestem = tarball[0:str(tarball).rfind(".tar.gz")]
-        tar = tarfile.open(tarball_path)
-        TarballFileNames = tar.getnames()
-        tar.close()
-    elif ".zip" in str(tarball):
-        outputnamestem = tarball[0:str(tarball).rfind(".zip")]
-        zip = zipfile.ZipFile(tarball_path)
-        TarballFileNames = zip.namelist()
-        zip.close()
-    else:
-        print("Invalid Tar file")
-
-    #Determining the configuration of the files and directories within the tarfile and defining variables for uniformity
-
-    FluxFileNameStems = []
-    extension = '.dat'
-    for IndividualFile in TarballFileNames:
-        extension_beginning = str(IndividualFile).rfind(".")
-        if extension_beginning > -1:
-            extension = str(IndividualFile[extension_beginning:])
-            ExtensionFound = str(IndividualFile).rfind(extension)
-            # gets rid of extension at the end of flux files
-            FluxFileNameStems.append(str(IndividualFile)[0:ExtensionFound])
-        else:
-            continue
-
-    smearvals = ["_smeared_w", "_smeared_u", "nts_w", "nts_u"]
-    #weightvals = ["_weighted", "unweighted"]
-
-    FilesToCleanup = []
-
-    #Add_funct sums up relevant files from output generated above
-    def add_funct(flux, detector, smear, *arg):
-        homebase = SNOwGLoBESdir + "/out/"
-
-        #Defining different variables for the combinatoric iterations
-        #And reformatting some variables for different naming uses
-        if smear == "_smeared_w" or smear == "nts_w":
-            weight_value = "weighted"
-        else:
-            weight_value = "unweighted"
-
-        if smear == "_smeared_w" or smear == "_smeared_u":
-            smear_value = "smeared"
-            smear_title = "detected"
[…]
