### 2021-11-08 f878eed
Adding collate test
