+++ requirements.txt
@@ -2,5 +2,6 @@ numpy >= 1.16
 scipy
 astropy
 pandas == 1.3.4
+Jinja2==3.0.2
 matplotlib
 h5py
