+++ requirements.txt
@@ -3,5 +3,6 @@ scipy
 astropy
 pandas == 1.3.4
 Jinja2==3.0.2
+tqdm
 matplotlib
 h5py
