### 2021-11-08 dcfb786
Integration test using the new interface
