### 2021-11-08 d10783c
Removing get_material from tests
