### 2021-11-09 699684e
Saving results in 'simulate'to hdf5  and reading them in 'collate'
