### 2021-11-09 5878617
Restored integration test
