### 2021-11-09 67c6efc
Backward-compatible format of returned values
