+++ python/snewpy/test/test_snowglobes.py
@@ -37,7 +37,6 @@ def sng():
 @pytest.mark.parametrize('detector, expected_total',crosscheck_table)
 def test_snowglobes_crosscheck(sng, detector, expected_total):
     flux = './models/Bollig_2016/fluence_Bollig_2016_s11.2c_AdiabaticMSW_NMO.dat'
-    material = get_material(detector)
     data = sng.run(flux,detector)
     total = data[0].weighted.smeared.sum().sum()
     assert total == pytest.approx(expected_total, 0.1)
@@ -46,7 +45,7 @@ def process(tarball_name):
     simulate(None,tarball_name,'icecube')
     collate(None, tarball_name,'icecube')
 
-@pytest.mark.timing    
+@pytest.mark.timing
 def test_simulation_chain_benchmark(benchmark):
     tarball_name='./models/Bollig_2016/fluence_Bollig_2016_s27.0c_AdiabaticMSW_IMO.tar.bz2'
     r = benchmark(process,tarball_name)
