### 2021-11-10 1ff7145
Fixed making plots for all combinations of smeared/weighted tables
