### 2021-11-10 ddbe245
Moving all models to models.ccsn module, also moving repeating get_init_spectra() to a new base class
