+++ python/snewpy/models/base.py
@@ -1,0 +1,258 @@
+from abc import abstractmethod, ABC
+
+from astropy.io import ascii, fits
+from astropy.table import Table, join
+from astropy.units.quantity import Quantity
+
+import numpy as np
+from scipy.interpolate import interp1d
+from scipy.special import loggamma, gamma, lpmv
+
+import os
+
+from snewpy.neutrino import Flavor
+from snewpy.flavor_transformation import *
+
+class SupernovaModel(ABC):
+    """Base class defining an interface to a supernova model."""
+    
+    def __init__(self):
+        pass
+
+    def __repr__(self):
+        if hasattr(self, "filename"):
+            # self.__class__ will be something like 
+            return f"{self.__class__.__name__}('{self.filename}')"
+        else:
+            return super().__repr__()
+
+    @abstractmethod
+    def get_time(self):
+        """Returns
+        -------
+            returns array of snapshot times from the simulation
+        """
+        pass
+
+    @abstractmethod
+    def get_initial_spectra(self, t, E, flavors=Flavor):
+        """Get neutrino spectra at the source.
+
+        Parameters
+        ----------
+        t : astropy.Quantity
+            Time to evaluate initial spectra.
+        E : astropy.Quantity or ndarray of astropy.Quantity
+            Energies to evaluate the initial spectra.
+        flavors: iterable of snewpy.neutrino.Flavor
+            Return spectra for these flavors only (default: all)
+
+        Returns
+        -------
+        initialspectra : dict
+            Dictionary of neutrino spectra, keyed by neutrino flavor.
+        """
+        pass
+
+    def get_initialspectra(self, *args):
+        """DO NOT USE! Only for backward compatibility!
+
+        :meta private:
+        """
+        print("[WARNING] Please use `get_initial_spectra()` instead of `get_initialspectra()`!")
+        return self.get_initial_spectra(*args)
+
+    def get_transformed_spectra(self, t, E, flavor_xform):
+        """Get neutrino spectra after applying oscillation.
+
+        Parameters
+        ----------
+        t : astropy.Quantity
+            Time to evaluate initial and oscillated spectra.
+        E : astropy.Quantity or ndarray of astropy.Quantity
+            Energies to evaluate the initial and oscillated spectra.
+        flavor_xform : FlavorTransformation
+            An instance from the flavor_transformation module.
+
+        Returns
+        -------
+        dict
+            Dictionary of transformed spectra, keyed by neutrino flavor.
+        """
+        initialspectra = self.get_initial_spectra(t, E)
+        transformed_spectra = {}
+
+        transformed_spectra[Flavor.NU_E] = \
+            flavor_xform.prob_ee(t, E) * initialspectra[Flavor.NU_E] + \
+            flavor_xform.prob_ex(t, E) * initialspectra[Flavor.NU_X]
+
+        transformed_spectra[Flavor.NU_X] = \
+            flavor_xform.prob_xe(t, E) * initialspectra[Flavor.NU_E] + \
+            flavor_xform.prob_xx(t, E) * initialspectra[Flavor.NU_X] 
+
+        transformed_spectra[Flavor.NU_E_BAR] = \
+            flavor_xform.prob_eebar(t, E) * initialspectra[Flavor.NU_E_BAR] + \
+            flavor_xform.prob_exbar(t, E) * initialspectra[Flavor.NU_X_BAR]
+
+        transformed_spectra[Flavor.NU_X_BAR] = \
+            flavor_xform.prob_xebar(t, E) * initialspectra[Flavor.NU_E_BAR] + \
+            flavor_xform.prob_xxbar(t, E) * initialspectra[Flavor.NU_X_BAR] 
+
+        return transformed_spectra   
+
+
[…]
