### 2021-11-10 0cdc2a6
Models test now use 'snewpy.models.ccsn'
