+++ python/snewpy/models/__init__.py
@@ -1,1 +1,1 @@
-import base,ccsn
+from . import base,ccsn
