### 2021-11-10 36fbd7f
Searching for model in `models.ccsn`
