### 2021-11-10 3872636
remove dependency on PyTables
