### 2021-11-10 f87ac09
Getting rid of .removesuffix to get compatibility with python <3.9
