### 2021-11-10 bf829fd
add pytest, pytest-benchmark as extras in setup.py Can now be installed using 'pip install .[dev]'
