### 2021-11-11 ae7f6c9
Docstrings to numpy format
