### 2021-11-11 0777abe
add pandas to doc requirements Needed because snewpy.snowglobes is imported to generate documentation from docstrings and the import fails if pandas isn’t installed
