### 2021-11-11 663db24
add packages to doc requirements Needed because snewpy.snowglobes is imported to generate documentation from docstrings and the import fails if packages aren’t installed
