### 2021-11-11 3a80cf2
Moving documentation requirements to extras_require
