+++ setup.py
@@ -57,6 +57,8 @@ with open('requirements.txt', 'r') as f:
         if line.strip():
             requires.append(line.strip())
 setup_keywords['install_requires'] = requires
+setup_keywords['docs'] = ['numpydoc']
+
 #
 # Internal data directories.
 #
