+++ doc/requirements.txt
@@ -1,6 +1,0 @@
-numpy
-scipy
-astropy
-matplotlib
-numpydoc
-h5py
