+++ setup.py
@@ -57,7 +57,9 @@ with open('requirements.txt', 'r') as f:
         if line.strip():
             requires.append(line.strip())
 setup_keywords['install_requires'] = requires
-setup_keywords['docs'] = ['numpydoc']
+setup_keywords['extras_require'] ={
+    'docs':['numpydoc']
+}
 
 #
 # Internal data directories.
