### 2021-11-11 21f91f2
Added test case for running in async loop (like in Jupyter)
