### 2021-11-12 972c3e2
Running asyncio loop in a new thread
