### 2021-11-12 af1b0d0
Returning exceptions, so if a particular run has failed, it doesn't crash all the others
