### 2021-11-12 f152b10
Smart progressbar to be shown in Jupyter"
