### 2021-11-15 #3772 @JostMigenda
👋 @PeterDenton — It’s been a couple of weeks since your last activity on this issue. If you got stuck while installing or using SNEWPY, can you please let us know? I’m happy to help to ensure that we can progress with this review!
