### 2021-11-16 0d07281
Adding stub for presuprnova
