### 2021-11-16 955a862
Fixed docstrings
