### 2021-11-16 96b34d0
Issue efficiency warning only once per detector, not for each flux file
