### 2021-11-16 1fb262e
Update README.md
Refer to notebooks included with downloaded model files
