### 2021-11-16 0dd3c32
minor docstring edits; remove unused import
