### 2021-11-16 a2a0118
Using `no_smear` option when there are no smearing files provided
