### 2021-11-16 82c2de5
Remove "d2O" experiment from the list, it's definition is incomplete
