### 2021-11-16 9906f54
delete unused variables from collate()
