### 2021-11-17 51cdb24
Remove forced use of Agg backend
