### 2021-11-17 9a5f6e4
Double energy bins for the  detector case
