### 2021-11-17 SNEWS2/snewpy#145 @PeterDenton
Problem downloading SN models: Warren a1.25 m120
When executing "python -c 'import snewpy; snewpy.get_models()'" many SN models download, but one fails:

Failed to download https://github.com/SNEWS2/snewpy/raw/v1.1b4/models/Warren_2020/stir_a1.25/stir_multimessenger_a1.25_m120.h5 to 'SNEWPY_models/Warren_2020/stir_a1.25/stir_multimessenger_a1.25_m120.h5'.

at the end it says:

ERROR: 1 exceptions occured. ([<HTTPError 502: 'Bad Gateway'>])

When I rerun it, it works fine and the 120M Warren model file is downloaded. Maybe it is hitting the server too fast? This isn't that important of an issue. Maybe you could add a comment if such this 502 error pops up that rerunning might help. Also, some files weren't attempted to download on the first pass and were only downloaded on the second pass.
