### 2021-11-17 SNEWS2/snewpy#146 @PeterDenton
GLoBES install version numbers in readme
In the readme there is a brief sketch on how to install GLoBES. I'm not sure if you only plan on interfacing with v3.2.17, but if newer version every come out (in fact they are up to 3.2.18 now) you may want to have a way to download the latest version. In any case, in the example you switch between 3.2.17 and 3.2.16.
