### 2021-11-17 SNEWS2/snewpy#147 @PeterDenton
text.usetex : True clashes with examples
I have "text.usetex : True" in my matplotlibrc file. In the examples there are strings of text which sometimes have underscores, derived from file names. While this is only for an example and is technically my fault, it probably isn't a super uncommon scenario. An easy fix is to put "mpl.rcParams['text.usetex'] = False" in the plotting examples if you want.
