### 2021-11-17 SNEWS2/snewpy#145 @JostMigenda
We’re intentionally limiting downloads to at most 8 files in parallel and any server-side rate limiting should return a 429 error (not 502) so I don’t think it’s related to that. I can’t reproduce it and since it worked for you on the 2nd try as well, it may have just been a blip in one of GitHub’s servers.

Good point about adding a comment though; I’ll do that.

> Also, some files weren't attempted to download on the first pass and were only downloaded on the second pass.

Hm … that should usually not happen.
`get_models()` will generally skip downloads if files are already present on disk; the only exception being the README files for each model directory which will always re-download. (That’s because we may add more progenitors to a model later—and update the README accordingly—whereas we don’t want to modify existing model files since the risk of confusion would be too great.)

So if it’s only the README files that were downloaded in the 2nd pass, that is intended. (Though they should still have been downloaded in the 1st pass—can you please check?) If it’s any other files, please let me know which specific ones so I can debug this.
