### 2021-11-17 SNEWS2/snewpy#145 @PeterDenton
Ah, it was just the readme files which were redownloaded and I see now that they were downloaded fine the first time, it's just a lot of lines and they all appeared at the end the second time.
