### 2021-11-17 SNEWS2/snewpy#146 @JostMigenda
I’ve just merged PR #112 that fixes the typo regarding 3.2.16 (and includes a few more updates to the README file). Thanks for the reminder!

With GLoBES 3.2.18, @schol mentioned that some SNOwGLoBES users had encountered problems so we recommend to stick with 3.2.17 for now. We may update that if/when there’s a new GLoBES release.
