### 2021-11-17 1e90ba1
improved error message if model file downloads fail
