### 2021-11-17 f7cb0e6
also redownload Jupyter notebooks
