+++ python/snewpy/models/__init__.py
@@ -1,1 +1,1 @@
-from . import base,ccsn
+from . import base,ccsn,presn
