### 2021-11-18 79ae1c3
Implemented Patton model interface
