+++ python/snewpy/models/presn.py
@@ -5,103 +5,132 @@ from astropy import units as u
 from .base import SupernovaModel
 from snewpy.neutrino import Flavor
 
-def _interpolate(x,y,z,new_x=None, new_y=None, logx=True):
-    if(new_x is not None): #interpolate in T
-        if(logx):
-            z = np.interp1d(np.log(x),np.log(z)
-                         ,axis=1, bounds_error=False, fill_value=0)(np.log(new_x))
-            z = np.exp(z)
-        else:
-            z = interp1d(x,z,axis=1, bounds_error=False, fill_value=0)(new_x)
-        x = new_x
-    if(new_y is not None): #interpolate in E
-        z = interp1d(y,z,axis=0, bounds_error=False, fill_value=0)(new_y)
-        y = new_y
-    return x,y,np.nan_to_num(z)
 
-def _interp_time(t0,v0, dt=1e-3):
+def _interp_T(t0, v0, dt=1e-3, dv=1e-10, axis=0):
     "dilog interpolation"
-    lt0 = np.log(t0+dt)
-    lv0 = np.log(v0)
-    lv1 = interp1d(lt0,lv0, axis=0)
-    def v1(t1):
-        return np.exp(lv1(np.log(t1+dt)))
-    return v1
+    lt0 = np.log(t0 + dt)
+    lv0 = np.log(v0 + dv)
+    lv1 = interp1d(lt0, lv0, axis=axis, fill_value=0, copy=False, bounds_error=False)
+    return lambda t1: np.exp(lv1(np.log(t1 + dt)))
+
+
+def _interp_E(e0, v0, axis=1):
+    "linear interpolation"
+    return interp1d(e0, v0, axis=axis, fill_value=0, copy=False, bounds_error=False)
+
+
+def _interp_TE(times, energies, array, ax_t=1, ax_e=2):
+    def _f(t, E):
+        a_t = _interp_T(times, array, axis=1)(t)
+        a_te = _interp_E(energies, a_t, axis=2)(E)
+        return a_te
+
+    return _f
+
 
 class Odrzywolek_2010(SupernovaModel):
     def __init__(self, fname):
-        df = pd.read_csv(fname,delim_whitespace=True, skiprows=1, usecols=[1,6,7,8],
-                     names=['time','a','alpha','b'], index_col='time')
-        #interpolated in time
-        self.df_t = _interp_time(df.index,df) 
+        df = pd.read_csv(
+            fname,
+            delim_whitespace=True,
+            skiprows=1,
+            usecols=[1,6,7,8],
+            names=["time","a","alpha","b"],
+            index_col="time",
+        )
+        # interpolated in time
+        self.df_t = _interp_T(df.index, df)
         self.times = df.index.to_numpy()
         self.factor = {}
         for f in Flavor:
             if f.is_electron:
-                self.factor[f]=1.
+                self.factor[f] = 1.0
             else:
-                #nuX/nuE ratio from Odrzywolek paper: (arXiv:astro-ph/0311012)
-                self.factor[f]=0.36 
-                
+                # nuX/nuE ratio from Odrzywolek paper: (arXiv:astro-ph/0311012)
+                self.factor[f] = 0.36
+
     def get_time(self):
-        return -self.times*u.s 
+        return -self.times * u.s
 
     def get_initial_spectra(self, t, E, flavors=Flavor):
-        #negative t for time before SN
-        t = -t.to_value('s')
-        E = E.to_value('MeV')
+        # negative t for time before SN
+        t = -t.to_value("s")
+        E = E.to_value("MeV")
         df = self.df_t(t)
-        a,alpha,b = df.T
-        Enu   = np.expand_dims(E,1)
[…]
