### 2021-11-19 c77ae73
Implemented Kato model interface
