### 2021-11-19 SNEWS2/snewpy#147 @JostMigenda
This sounds like a reasonable request to me. But before we go ahead and manually edit every single one of the many example notebooks and scripts included in SNEWPY, it’s probably worth taking a moment to think: Are there any other common settings that we should take into account? (And if so, is it worth creating a [custom style sheet](https://matplotlib.org/stable/tutorials/introductory/customizing.html#using-style-sheets) for SNEWPY?)
