### 2021-11-19 #3772 @JostMigenda
For posterity, since I forgot to mention this JOSS review issue in the respective issues:
* SNEWS2/snewpy#145 and SNEWS2/snewpy#146 were opened as part of the review and are fixed already.
* SNEWS2/snewpy#147 is still open, but more of a suggestion so it shouldn’t block progress here—@PeterDenton, I hope I’m not misrepresenting you here!

It looks like both review checklists are completed now. I’ll check if Whedon is happy with the paper.
