### 2021-11-19 SNEWS2/snewpy#147 @PeterDenton
I thought about a custom style sheet too, but I only suggested turning usetex off because that was the only one that broke the examples that I could think of. My other things might change how the plots look, but at least they would be generated. But trying to write a string "asdf_jkl" without dollar signs around it will make matplotlib a bit upset and throw a bunch of latex garbage in the terminal.
