### 2021-11-19 0287526
Using DeprecationWarnings instead of print
