### 2021-11-20 3db4000
Adding function to calculate flux at certain distance
