### 2021-11-20 9d5c035
Adding function to calculate flux at certain distance
