### 2021-11-20 3890821
Using new function and vectorized calculation
