+++ python/snewpy/snowglobes.py
@@ -34,13 +34,14 @@ import matplotlib as mpl
 import matplotlib.pyplot as plt
 import numpy as np
 from astropy import units as u
+import pandas as pd
+from tempfile import TemporaryDirectory
 
 import snewpy.models
 from snewpy.flavor_transformation import *
 from snewpy.neutrino import Flavor, MassHierarchy
-
-mpl.use('Agg')
-
+from scipy.integrate import cumulative_trapezoid
+from pathlib import Path
 
 def generate_time_series(model_path, model_type, transformation_type, d, output_filename=None, ntbins=30, deltat=None):
     """Generate time series files in SNOwGLoBES format.
@@ -76,74 +77,69 @@ def generate_time_series(model_path, model_type, transformation_type, d, output_
     flavor_transformation_dict = {'NoTransformation': NoTransformation(), 'AdiabaticMSW_NMO': AdiabaticMSW(mh=MassHierarchy.NORMAL), 'AdiabaticMSW_IMO': AdiabaticMSW(mh=MassHierarchy.INVERTED), 'NonAdiabaticMSWH_NMO': NonAdiabaticMSWH(mh=MassHierarchy.NORMAL), 'NonAdiabaticMSWH_IMO': NonAdiabaticMSWH(mh=MassHierarchy.INVERTED), 'TwoFlavorDecoherence': TwoFlavorDecoherence(), 'ThreeFlavorDecoherence': ThreeFlavorDecoherence(), 'NeutrinoDecay_NMO': NeutrinoDecay(mh=MassHierarchy.NORMAL), 'NeutrinoDecay_IMO': NeutrinoDecay(mh=MassHierarchy.INVERTED)}
     flavor_transformation = flavor_transformation_dict[transformation_type]
 
-    model_dir, model_file = os.path.split(os.path.abspath(model_path))
+    model_path = Path(model_path)
     snmodel = model_class(model_path)
 
     # Subsample the model time. Default to 30 time slices.
-    tmin = snmodel.get_time()[0]
-    tmax = snmodel.get_time()[-1]
+    tmin = snmodel.get_time()[0].to_value('s')
+    tmax = snmodel.get_time()[-1].to_value('s')
     if deltat is not None:
-        dt = deltat
-        ntbins = int((tmax-tmin)/dt)
+        tedges = np.arange(tmin, tmax, dt)
     else:
-        dt = (tmax - tmin) / (ntbins+1)
+        tedges = np.linspace(tmin,tmax,ntbins+1)
 
-    tedges = np.arange(tmin/u.s, tmax/u.s, dt/u.s)*u.s
-    times = 0.5*(tedges[1:] + tedges[:-1])
+    times = 0.5*(tedges[1:] + tedges[:-1])<<u.s
+    dt = np.diff(tedges)
 
-    # Generate output.
-    if output_filename is not None:
-        tfname = output_filename + 'kpc.tar.bz2'
-    else:
-        model_file_root, _ = os.path.splitext(model_file)  # strip extension (if present)
-        tfname = model_file_root + '.' + transformation_type + '.{:.3f},{:.3f},{:d}-{:.1f}'.format(tmin, tmax, ntbins, d) + 'kpc.tar.bz2'
-
-    with tarfile.open(os.path.join(model_dir, tfname), 'w:bz2') as tf:
+    energy = np.linspace(0, 100, 501) * u.MeV
+    osc_fluence = snmodel.get_transformed_flux(times,energy,flavor_transformation, d*u.kpc)
+    #construct a 3D array (Flavors * Time * Energy)
+    osc_fluence_array = np.stack([osc_fluence[f] for f in sorted(osc_fluence)],axis=0)
[…]
