### 2021-11-20 a0b59c8
Remove NaN values
