+++ python/snewpy/models/base.py
@@ -286,9 +286,12 @@ class PinchedModel(SupernovaModel):
             Ea = np.expand_dims(Ea,axis=1)
             a  = np.expand_dims(a, axis=1)
             # For numerical stability, evaluate log PDF and then exponentiate.
-            initialspectra[flavor] = \
+            result = \
               np.exp(np.log(L) - (2+a)*np.log(Ea) + (1+a)*np.log(1+a)
                     - loggamma(1+a) + a*np.log(E) - (1+a)*(E/Ea)) / (u.erg * u.s)
+            #remove bad values
+            result[np.isnan(result)] = 0
+            initialspectra[flavor] = result
 
         return initialspectra
 
