### 2021-11-20 d5e3969
Remove NaN values
