### 2021-11-20 e7a2304
Integrating the flux in time and energy bins
