### 2021-11-20 148e3ba
Moving the common code to the small helper functions
