### 2021-11-20 474465f
Reworked generate_fluence
