+++ python/snewpy/snowglobes.py
@@ -42,16 +42,22 @@ from snewpy.flavor_transformation import *
 from snewpy.neutrino import Flavor, MassHierarchy
 from pathlib import Path
 from scipy.integrate import cumulative_trapezoid
+from scipy.interpolate import interp1d
 from contextlib import contextmanager
 
-def _integrate_bins(bin_edges, array, axis):
+def _cumulative_integral(bin_edges, array, axis):
     """Integrate the array along an axis"""
-    cumtrapz = cumulative_trapezoid(
+    return  cumulative_trapezoid(
         array, x=bin_edges,
         initial = 0,
         axis=axis
     )
-    result = np.diff(cumtrapz,axis=axis)
+
+
+def _integrate_bins(bin_edges, array, axis):
+    """Integrate the array along an axis"""
+    cumtrapz= _cumulative_integral(bin_edges, array, axis)
+    result = np.diff(cumtrapz, axis=axis)
     return result
 
 
@@ -67,7 +73,7 @@ def _save_flux_table(energy, flux, filename, t, dt):
 
     header = f'TBinMid={t:g}sec TBinWidth={dt:g}s EBinWidth=0.2MeV Fluence at Earth for this timebin in neutrinos per cm^2\n'
     header+=' '.join(f'{key:>16}' for key in table)
-# Generate energy + number flux table.
+    # Generate energy + number flux table.
     table = np.stack(list(table.values()))
     np.savetxt(filename, table.T, header=header, fmt='%17.8E',delimiter='')
 
@@ -140,14 +146,15 @@ def generate_time_series(model_path, model_type, transformation_type, d, output_
     dt = np.diff(tedges)
 
     energy = np.linspace(0, 100, 501) * u.MeV
-    osc_fluence = snmodel.get_transformed_flux(tedges,energy,flavor_transformation, d*u.kpc)
+    flux = snmodel.get_transformed_flux(tedges,energy,flavor_transformation, d*u.kpc)
     #construct a 3D array (Flavors * Time * Energy)
-    osc_fluence_array = np.stack([osc_fluence[f] for f in sorted(osc_fluence)],axis=0)
+    flux_array = np.stack([flux[f] for f in sorted(flux)],axis=0)
     #get rid of the units
-    osc_fluence_array = osc_fluence_array.to_value('1/(MeV*s*cm**2)')
+    flux_array = flux_array.to_value('1/(MeV*s*cm**2)')
     #integrate the flux in each energy bin
-    osc_fluence_array=_integrate_bins(energy,osc_fluence_array,axis=2)
-    osc_fluence_array=_integrate_bins(tedges,osc_fluence_array,axis=1)
+    flux_array=_integrate_bins(energy,flux_array,axis=2)
+    #integrate the flux in each time bin
+    flux_array=_integrate_bins(tedges,flux_array,axis=1)
     # Save all to tar file
     if output_filename is None:
         output_filename = f'{model_path.stem}.{transformation_type}.{tmin:.3f} s,{tmax:.3f} s,{ntbins:d}-{d:.1f}'
@@ -161,10 +168,10 @@ def generate_time_series(model_path, model_type, transformation_type, d, output_
         energy = energy[:-1].to_value('GeV')
         # Loop over sampled times.
         for i, t in enumerate(times):
-            osc_fluence = osc_fluence_array[:,i,:]
+            fluence = flux_array[:,i,:]
             filename = f'{model_path.stem}.tbin{i+1:01d}.{transformation_type}'+\
                        f'.{tmin:.3f},{tmax:.3f},{ntbins:01d}-{d:.1f}kpc.dat'
-            _save_flux_table(energy,osc_fluence,tempdir/filename,t,dt[i])
+            _save_flux_table(energy,fluence,tempdir/filename,t,dt[i])
 
     return output_filename
 
[…]
