### 2021-11-21 431136d
Using _integrate_bins0 function to multiply by bin width
