### 2021-11-21 7b7b9a6
Moving __str__ and _repr_markdown_ to the base class
