+++ python/snewpy/models/ccsn.py
@@ -77,18 +77,6 @@ class Analytic3Species(PinchedModel):
             self.meanE[flavor] = simtab['E_{}'.format(_flav.name)] * u.MeV
             self.pinch[flavor] = simtab['ALPHA_{}'.format(_flav.name)]
 
-    def __str__(self):
-        """Default representation of the model.
-        """
-        mod = 'Analytic Model: {}\n'.format(self.filename)
-        return mod
-
-    def _repr_markdown_(self):
-        """Markdown representation of the model, for Jupyter notebooks.
-        """
-        mod = '**Analytic Model**: {}\n\n'.format(self.filename)
-        return mod  
-
 
 class Nakazato_2013(PinchedModel):
     """Set up a model based on simulations from Nakazato et al., ApJ S 205:2
@@ -116,6 +104,12 @@ class Nakazato_2013(PinchedModel):
             self.revival_time = 0 * u.ms
             self.EOS = filename.split('-')[-4].upper()
 
+        self.metadata = {
+            'Progenitor mass':self.progenitor_mass,
+            'EOS':self.EOS,
+            'Metallicity':self.metallicity,
+            'Revival time':self.revival_time
+            }
         # Read FITS table using the astropy reader.
         simtab = Table.read(filename)
         self.filename = os.path.basename(filename)
@@ -138,30 +132,6 @@ class Nakazato_2013(PinchedModel):
             self.meanE[flavor] = simtab['E_{}'.format(_flav.name)].to('MeV')
             self.pinch[flavor] = simtab['ALPHA_{}'.format(_flav.name)]
 
-    def __str__(self):
-        """Default representation of the model.
-        """
-        mod = 'Nakazato_2013 Model: {}\n'.format(self.filename)
-        s = ['Progenitor mass : {}'.format(self.progenitor_mass),
-             'Metallicity     : {}'.format(self.metallicity),
-             'Revival time    : {}'.format(self.revival_time),
-             'Eq. of state    : {}'.format(self.EOS)
-             ]
-        return mod + '\n'.join(s)
-        
-    def _repr_markdown_(self):
-        """Markdown representation of the model, for Jupyter notebooks.
-        """
-        mod = '**Nakazato_2013 Model**: {}\n\n'.format(self.filename)
-        s = ['|Parameter|Value|',
-             '|:---------|:-----:|',
-             '|Progenitor mass | ${0.value:g}$ {0.unit:latex}|'.format(self.progenitor_mass),
-             '|Metallicity | ${:g}$|'.format(self.metallicity),
-             '|Revival time | ${0.value:g}$ {0.unit:latex}|'.format(self.revival_time),
-             '|EOS | {}|'.format(self.EOS)
-             ]
-        return mod + '\n'.join(s)
-
 
 class Sukhbold_2015(PinchedModel):
     """Set up a model based on simulations from Sukhbold et al., ApJ 821:38,2016. Models were shared privately by email.
@@ -178,6 +148,11 @@ class Sukhbold_2015(PinchedModel):
         # Store model metadata.
         self.progenitor_mass = float(filename.split('-')[-1].strip('z%.fits')) * u.Msun
         self.EOS = filename.split('-')[-2]
+
+        self.metadata = {
+            'Progenitor mass':self.progenitor_mass,
+            'EOS':self.EOS,
+            }
 
         # Read FITS table using the astropy unified Table reader.
         simtab = Table.read(filename)
@@ -196,103 +171,24 @@ class Sukhbold_2015(PinchedModel):
             self.luminosity[flavor] = simtab['L_{}'.format(flavor.name)].to('erg/s')
[…]
