### 2021-11-21 bf1dffd
Moving __str__ and _repr_markdown_ to the base class
