### 2021-11-21 e645821
Depracation warning when importing ccsn models directly from snewpy.models
