### 2021-11-21 364f15d
Depracation warning when importing ccsn models directly from snewpy.models
