### 2021-11-22 #3772 @dfm
@JostMigenda: Thanks for your patience. I've checked in about the formatting issue with the manuscript and we're actually just going to fix that in post production, so for now, we're going to proceed with the final processing. Here are the steps that I'll need from you:

Thanks!
