### 2021-11-22 #3772 @JostMigenda
@dfm I’ve tagged a new version (`v1.1`) and archived it to Zenodo ([DOI:10.5281/zenodo.5719209](https://zenodo.org/record/5719209)). The metadata that’s automatically generated by GitHub isn’t correct; I’ll contact a co-author who has permissions to edit that and will respond here as soon as that’s fixed.
