### 2021-11-22 9b22855
Using proper methods and FutureWarnings
