### 2021-11-22 e8c49b0
Using FutureWarning instead of UserWarning
