+++ .github/workflows/integration.yml
@@ -25,14 +25,44 @@ jobs:
         uses: actions/setup-python@v2
         with:
           python-version: ${{ matrix.python-version }}
+
+      - name: Caching requirements
+      - uses: actions/cache@v2
+      - id:   cache_pip
+        with:
+          path: ~/.cache/pip
+          key: ${{ runner.os }}-pip-${{ steps.setup-python.outputs.python-version }}-${{ hashFiles('**/requirements.txt') }}
+
       - name: Install dependencies
+        if: steps.cache_pip.output.cache-hit != 'true'
         run: |
           python -m pip install --upgrade pip
           if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
-      - name: Install SNEWPY
+
+      - name: Caching GLoBES
+      - uses: actions/cache@v2
+        id: cache_globes
+        with:
+          path: |
+            ${GITHUB_WORKSPACE}/opt/globes-3.2.17-install
+          key: ${{ runner.os }}-globes
+
+      - name: Check latest SNOwGLoBES tag
+        id: snowglobes_tag
         run: |
-          python setup.py install --user
-      - name: Install SNOwGLoBES
+          tag=$(git ls-remote --sort='version:refname' --tags --refs https://github.com/SNOwGLoBES/snowglobes.git | tail -1 | cut -d'/' -f3)
+          echo '::set-output name=latest::${tag}'
+
+      - name: Caching SNOwGLoBES
+      - uses: actions/cache@v2
+        id: cache_snowglobes
+        with:
+          path: |
+            ${GITHUB_WORKSPACE}/opt/globes-3.2.17-install
+            ${GITHUB_WORKSPACE}/opt/snowglobes
+          key: ${{ runner.os }}-snowglobes-${{ steps.snowglobes_tag.output.latest }}
+      - name: Install GLoBES
+        if: steps.cache_globes.outputs.cache-hit != 'true'
         run: |
           mkdir opt
           cd opt
@@ -42,6 +72,10 @@ jobs:
           ./configure --prefix=$GITHUB_WORKSPACE/opt/globes-3.2.17-install  --disable-binary
           make
           make install
+
+      - name: Install SNOwGLoBES
+        if: steps.cache_snowglobes.outputs.cache-hit != 'true'
+        run: |
           export GLB_DIR=${GITHUB_WORKSPACE}/opt/globes-3.2.17-install
           # Install SNOwGLoBES
           cd ${GITHUB_WORKSPACE}/opt/
@@ -51,6 +85,11 @@ jobs:
           cd src
           make
           make install
+
+      - name: Install SNEWPY
+        run: |
+          python setup.py install --user
+
       - name: Run Integration Test
         run: |
           python -m unittest python/snewpy/test/snowglobes_integrationtest.py
