+++ .github/workflows/tests.yml
@@ -31,10 +31,18 @@ jobs:
         uses: actions/setup-python@v2
         with:
           python-version: ${{ matrix.python-version }}
+
+      - name: Caching requirements
+      - uses: actions/cache@v2
+      - id:   cache_pip
+        with:
+          path: ~/.cache/pip
+          key: ${{ runner.os }}-pip-${{ steps.setup-python.outputs.python-version }}-${{ hashFiles('**/requirements.txt') }}
+
       - name: Install dependencies
+        if: steps.cache_pip.output.cache-hit != 'true'
         run: |
           python -m pip install --upgrade pip
-          pip install pytest
           if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
       - name: Install SNEWPY
         run: |
