+++ .github/workflows/integration.yml
@@ -27,8 +27,8 @@ jobs:
           python-version: ${{ matrix.python-version }}
 
       - name: Caching requirements
-      - uses: actions/cache@v2
-      - id:   cache_pip
+        uses: actions/cache@v2
+        id:   cache_pip
         with:
           path: ~/.cache/pip
           key: ${{ runner.os }}-pip-${{ steps.setup-python.outputs.python-version }}-${{ hashFiles('**/requirements.txt') }}
@@ -40,7 +40,7 @@ jobs:
           if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
 
       - name: Caching GLoBES
-      - uses: actions/cache@v2
+        uses: actions/cache@v2
         id: cache_globes
         with:
           path: |
@@ -54,7 +54,7 @@ jobs:
           echo '::set-output name=latest::${tag}'
 
       - name: Caching SNOwGLoBES
-      - uses: actions/cache@v2
+        uses: actions/cache@v2
         id: cache_snowglobes
         with:
           path: |
