+++ .github/workflows/tests.yml
@@ -33,8 +33,8 @@ jobs:
           python-version: ${{ matrix.python-version }}
 
       - name: Caching requirements
-      - uses: actions/cache@v2
-      - id:   cache_pip
+        uses: actions/cache@v2
+        id:   cache_pip
         with:
           path: ~/.cache/pip
           key: ${{ runner.os }}-pip-${{ steps.setup-python.outputs.python-version }}-${{ hashFiles('**/requirements.txt') }}
