+++ .github/workflows/tests.yml
@@ -43,6 +43,7 @@ jobs:
         if: steps.cache_pip.output.cache-hit != 'true'
         run: |
           python -m pip install --upgrade pip
+          python -m pip install pytest
           if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
       - name: Install SNEWPY
         run: |
