### 2021-11-22 b27c230
Fixing command retrieving tag
