### 2021-11-22 0a1254e
Using only the latest tag of SNOwGLoBES
