### 2021-11-25 a45a5eb
SupernovaModel now can produce Flux object directly
