### 2021-11-25 7f86190
Fixing the shape of output arrays
