### 2021-11-25 d9eebf6
Using Flux class in generate_fluence/time_series
