### 2021-11-25 29dbe1e
run tests under all supported Python versions
